Update Editor-Area component and Editor Route/Controller to notify/receive when editor content has changed.  The process is now decoupled from the model.body property, and instead sends an action up to the controller that will modify a '_contentDidChange' property to signal a change in HTML content rather than modify the model.body property directly which had negative effects when the data was persisted and returned from API server.
The update also has a function on the component, when a render is completed (afterRender) to inject the model.body content into the editor-area component or take the editor-area as is if there is no content in model.body (null or undefined or blank...).

TODO: Update the save method on the editor controller to do any sanitizing (if any) before a model.save() is called.  Likely not necessary since we'll be doing this on the server side anyways, and the returned data from a save is not being reinjected into the editor.